DSEGOG-174 plot colours are now set from plot config state
Previously, these were reset whenever the plot was opened. Now they can be restored properly
