+++ src/plotting/plotSettings/colourGenerator.test.tsx
@@ -1,0 +1,189 @@
+import ColourGenerator from './colourGenerator';
+
+describe('ColourGenerator', () => {
+  let colourGenerator;
+  const colourOrder = [
+    '#008000', // dark green
+    '#0000ff', // dark blue
+    '#ff00ff', // pink
+    '#00ffff', // light blue
+    '#008080', // teal
+    '#800000', // deep red
+    '#00ff00', // light green
+    '#000080', // navy blue
+    '#7f8000', // brown-ish yellow?
+    '#80007f', // indigo
+  ];
+
+  beforeEach(() => {
+    colourGenerator = new ColourGenerator();
+  });
+
+  describe('getting next colour', () => {
+    it('returns the next colour in the remaining colours list', () => {
+      const colour = colourGenerator.nextColour();
+      expect(colour).toEqual(colourOrder[0]);
+    });
+
+    it('returns a random colour if the list of remaining colours is empty', () => {
+      colourOrder.forEach(() => {
+        colourGenerator.nextColour();
+      });
+      const colour = colourGenerator.nextColour();
+      expect(colourOrder.includes(colour)).toBeFalsy();
+    });
+  });
+
+  describe('removing a colour', () => {
+    it('1 colour selected', () => {
+      const firstGenerated = colourGenerator.nextColour();
+
+      colourGenerator.removeColour(firstGenerated);
+
+      const secondGenerated = colourGenerator.nextColour();
+      expect(firstGenerated).toEqual(secondGenerated);
+    });
+
+    describe('2 colours selected', () => {
+      it('removes first', () => {
+        const firstGenerated = colourGenerator.nextColour();
+        colourGenerator.nextColour();
+
+        colourGenerator.removeColour(firstGenerated);
+
+        const thirdGenerated = colourGenerator.nextColour();
+        expect(thirdGenerated).toEqual(firstGenerated);
+      });
+
+      it('removes second', () => {
+        colourGenerator.nextColour();
+        const secondGenerated = colourGenerator.nextColour();
+
+        colourGenerator.removeColour(secondGenerated);
+
+        const thirdGenerated = colourGenerator.nextColour();
+        expect(thirdGenerated).toEqual(secondGenerated);
+      });
+    });
+
+    describe('3 colours selected', () => {
+      it('removes first', () => {
+        const firstGenerated = colourGenerator.nextColour();
+        colourGenerator.nextColour();
+        colourGenerator.nextColour();
+
+        colourGenerator.removeColour(firstGenerated);
+
+        const fourthGenerated = colourGenerator.nextColour();
+        expect(fourthGenerated).toEqual(firstGenerated);
+      });
+
+      it('removes middle', () => {
+        colourGenerator.nextColour();
+        const secondGenerated = colourGenerator.nextColour();
+        colourGenerator.nextColour();
+
+        colourGenerator.removeColour(secondGenerated);
+
+        const fourthGenerated = colourGenerator.nextColour();
+        expect(fourthGenerated).toEqual(secondGenerated);
+      });
+
+      it('removes last', () => {
+        colourGenerator.nextColour();
+        colourGenerator.nextColour();
+        const thirdGenerated = colourGenerator.nextColour();
+
+        colourGenerator.removeColour(thirdGenerated);
+
+        const fourthGenerated = colourGenerator.nextColour();
+        expect(fourthGenerated).toEqual(thirdGenerated);
+      });
+    });
+
+    describe('10 colours selected', () => {
+      it('removes first', () => {
+        const firstGenerated = colourGenerator.nextColour();
+        colourGenerator.nextColour();
+        colourGenerator.nextColour();
+        colourGenerator.nextColour();
+        colourGenerator.nextColour();
+        colourGenerator.nextColour();
+        colourGenerator.nextColour();
+        colourGenerator.nextColour();
+        colourGenerator.nextColour();
+        colourGenerator.nextColour();
+
+        colourGenerator.removeColour(firstGenerated);
+
+        const fourthGenerated = colourGenerator.nextColour();
+        expect(fourthGenerated).toEqual(firstGenerated);
+      });
+
+      it('removes some middle ones', () => {
+        colourGenerator.nextColour();
+        colourGenerator.nextColour();
+        const thirdGenerated = colourGenerator.nextColour();
+        colourGenerator.nextColour();
+        const fifthGenerated = colourGenerator.nextColour();
+        colourGenerator.nextColour();
+        colourGenerator.nextColour();
+        colourGenerator.nextColour();
+        const ninthGenerated = colourGenerator.nextColour();
+        colourGenerator.nextColour();
+
+        // Remove third, then ninth, then fifth
+        colourGenerator.removeColour(thirdGenerated);
+        colourGenerator.removeColour(ninthGenerated);
+        colourGenerator.removeColour(fifthGenerated);
+
+        const eleventhGenerated = colourGenerator.nextColour();
+        const twelfthGenerated = colourGenerator.nextColour();
+        const thirteenthGenerated = colourGenerator.nextColour();
+        expect(eleventhGenerated).toEqual(thirdGenerated);
+        expect(twelfthGenerated).toEqual(fifthGenerated);
+        expect(thirteenthGenerated).toEqual(ninthGenerated);
+      });
+
+      it('removes last', () => {
+        colourGenerator.nextColour();
+        colourGenerator.nextColour();
+        colourGenerator.nextColour();
+        colourGenerator.nextColour();
+        colourGenerator.nextColour();
+        colourGenerator.nextColour();
+        colourGenerator.nextColour();
+        colourGenerator.nextColour();
+        colourGenerator.nextColour();
+        const tenthGenerated = colourGenerator.nextColour();
+
+        colourGenerator.removeColour(tenthGenerated);
+
+        const eleventhGenerated = colourGenerator.nextColour();
+        expect(eleventhGenerated).toEqual(tenthGenerated);
+      });
+    });
+
+    describe('11 colours selected', () => {
+      it('removes last', () => {
+        colourGenerator.nextColour();
+        colourGenerator.nextColour();
+        colourGenerator.nextColour();
+        colourGenerator.nextColour();
+        colourGenerator.nextColour();
+        colourGenerator.nextColour();
+        colourGenerator.nextColour();
+        colourGenerator.nextColour();
+        colourGenerator.nextColour();
+        colourGenerator.nextColour();
+        const eleventhGenerated = colourGenerator.nextColour();
+
+        colourGenerator.removeColour(eleventhGenerated);
+
+        // No more colours in remaining colours list, so we expect a random colour
+        const twelfthGenerated = colourGenerator.nextColour();
+        expect(twelfthGenerated).not.toEqual(eleventhGenerated);
+      });
+    });
+  });
+});
