+++ src/plotting/util.tsx
@@ -7,10 +7,7 @@ export const deepCopySelectedPlotChannels = (
   return original.map((currentChannel: SelectedPlotChannel) => ({
     name: currentChannel.name,
     options: {
-      visible: currentChannel.options.visible,
-      lineStyle: currentChannel.options.lineStyle,
-      colour: currentChannel.options.colour,
-      yAxis: currentChannel.options.yAxis,
+      ...currentChannel.options,
     },
   }));
 };
